+++ TwitterImagePipeline/Project/TIPImageDownloadInternalContext.m
@@ -15,9 +15,6 @@
 NS_ASSUME_NONNULL_BEGIN
 
 @implementation TIPImageDownloadInternalContext
-{
-    NSMutableArray<id<TIPImageDownloadDelegate>> *_delegates;
-}
 
 - (instancetype)init
 {
@@ -73,7 +70,7 @@ NS_ASSUME_NONNULL_BEGIN
     NSUInteger count = _delegates.count;
     [_delegates removeObject:delegate];
     if (count > _delegates.count) {
-        id<TIPImageFetchDownload> download = self.download;
+        id<TIPImageFetchDownload> download = _download;
         [TIPImageDownloadInternalContext executeDelegate:delegate suspendingQueue:NULL block:^(id<TIPImageDownloadDelegate> blockDelegate) {
             [blockDelegate imageDownload:(id)download didCompleteWithPartialImage:nil lastModified:nil byteSize:0 imageType:nil image:nil imageRenderLatency:0.0 statusCode:0 error:[NSError errorWithDomain:TIPImageFetchErrorDomain code:TIPImageFetchErrorCodeCancelled userInfo:nil]];
         }];
@@ -81,7 +78,7 @@ NS_ASSUME_NONNULL_BEGIN
 }
 
 - (void)executePerDelegateSuspendingQueue:(nullable dispatch_queue_t)queue
-                                    block:(void(^)(id<TIPImageDownloadDelegate>))block;
+                                    block:(void(^)(id<TIPImageDownloadDelegate>))block
 {
     for (id<TIPImageDownloadDelegate> delegate in _delegates) {
         [TIPImageDownloadInternalContext executeDelegate:delegate
@@ -91,14 +88,14 @@ NS_ASSUME_NONNULL_BEGIN
 
 + (void)executeDelegate:(id<TIPImageDownloadDelegate>)delegate
         suspendingQueue:(nullable dispatch_queue_t)queue
-                  block:(void (^)(id<TIPImageDownloadDelegate>))block;
+                  block:(void (^)(id<TIPImageDownloadDelegate>))block
 {
     dispatch_queue_t delegateQueue = [delegate respondsToSelector:@selector(imageDownloadDelegateQueue)] ? delegate.imageDownloadDelegateQueue : NULL;
     if (delegateQueue) {
         if (queue) {
             dispatch_suspend(queue);
         }
-        dispatch_async(delegateQueue, ^{
+        tip_dispatch_async_autoreleasing(delegateQueue, ^{
             block(delegate);
             if (queue) {
                 dispatch_resume(queue);