+++ TwitterImagePipeline/Project/TIPLRUCache.m
@@ -316,4 +316,3 @@ NS_INLINE void TIPLRUCacheAssertHeadAndTail(TIPLRUCache *cache)
 @end
 
 NS_ASSUME_NONNULL_END
-