+++ TwitterImagePipelineTests/TIPTestURLProtocol.h
@@ -46,4 +46,3 @@ FOUNDATION_EXTERN NSString * const TIPTestURLProtocolErrorDomain;
 @end
 
 NS_ASSUME_NONNULL_END
-